% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{apply_imputed_glucose}
\alias{apply_imputed_glucose}
\title{Apply imputed glucose values to standardized CGM data}
\usage{
apply_imputed_glucose(data, imputation_result, model = "mice_only")
}
\arguments{
\item{data}{Standardized CGM data.}

\item{imputation_result}{Result from \code{run_cgmissingdata_imputation()}.}

\item{model}{Imputation model name.}
}
\value{
Standardized CGM data with imputed glucose values and flags.
}
\description{
Apply imputed glucose values to standardized CGM data
}
