% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGManalyzer2-package.R
\docType{package}
\name{CGManalyzer2-package}
\alias{CGManalyzer2}
\alias{CGManalyzer2-package}
\title{CGManalyzer2}
\description{
Modular Shiny application for continuous glucose monitoring analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ZhangLabUKY/CGMAnalyzerShiny}
  \item Report bugs at \url{https://github.com/ZhangLabUKY/CGMAnalyzerShiny/issues}
}

}
\author{
\strong{Maintainer}: Shubh Saraswat \email{shubh.saraswat00@gmail.com} (\href{https://orcid.org/0009-0009-2359-1484}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Xiaohua Douglas Zhang \email{douglas.zhang@uky.edu} (\href{https://orcid.org/0000-0002-2486-7931}{ORCID})
}

}
\keyword{internal}
