% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGMAnalyzerShiny-package.R
\docType{package}
\name{CGMAnalyzerShiny-package}
\alias{CGMAnalyzerShiny}
\alias{CGMAnalyzerShiny-package}
\title{CGMAnalyzerShiny}
\description{
Modular Shiny application for continuous glucose monitoring analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ZhangLabUKY/CGMAnalyzerShiny}
  \item Report bugs at \url{https://github.com/ZhangLabUKY/CGMAnalyzerShiny/issues}
}

}
\author{
\strong{Maintainer}: Shubh Saraswat \email{shubh.saraswat00@gmail.com}

}
\keyword{internal}
