% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{cgmissingdata_available}
\alias{cgmissingdata_available}
\title{Check whether CGMissingDataR imputation is available}
\usage{
cgmissingdata_available()
}
\value{
Logical.
}
\description{
Check whether CGMissingDataR imputation is available
}
