% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_agp_summary_plot}
\alias{create_agp_summary_plot}
\title{Create an AGP-style CGM summary plot}
\usage{
create_agp_summary_plot(
  data,
  thresholds = default_cgm_thresholds(),
  participant = "",
  group = "",
  visit = "",
  bin_minutes = 30
)
}
\arguments{
\item{data}{Standardized CGM data.}

\item{thresholds}{Named glucose thresholds in mg/dL.}

\item{participant}{Subject ID filter.}

\item{group}{Group filter.}

\item{visit}{Visit filter.}

\item{bin_minutes}{Time-of-day bin width in minutes.}
}
\value{
A ggplot object.
}
\description{
Create an AGP-style CGM summary plot
}
