% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness.R
\name{create_missingness_timeline_plot}
\alias{create_missingness_timeline_plot}
\title{Create a missingness timeline plot}
\usage{
create_missingness_timeline_plot(data)
}
\arguments{
\item{data}{Standardized CGM data.}
}
\value{
A ggplot object.
}
\description{
Create a missingness timeline plot
}
