% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_tir_plot}
\alias{create_tir_plot}
\title{Create a TIR stacked bar plot}
\usage{
create_tir_plot(metrics)
}
\arguments{
\item{metrics}{Metric summary data frame from \code{compute_core_metrics()}.}
}
\value{
A ggplot object.
}
\description{
Create a TIR stacked bar plot
}
