% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization.R
\name{load_demo_cgm_data}
\alias{load_demo_cgm_data}
\title{Load bundled demo CGM data}
\usage{
load_demo_cgm_data()
}
\value{
A data frame with demo CGM readings.
}
\description{
Load bundled demo CGM data
}
