% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_standardization.R
\name{parse_cgm_timestamp}
\alias{parse_cgm_timestamp}
\title{Parse CGM timestamps}
\usage{
parse_cgm_timestamp(x, tz = "UTC", date_order = "dmy")
}
\arguments{
\item{x}{Timestamp vector.}

\item{tz}{Time zone used when timestamps do not include one.}

\item{date_order}{Date order for non-year-first dates. Defaults to
day-first for CGM device exports.}
}
\value{
POSIXct vector.
}
\description{
Parse CGM timestamps
}
