% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_display.R
\name{prepare_metrics_display}
\alias{prepare_metrics_display}
\title{Prepare metrics for user-facing display}
\usage{
prepare_metrics_display(metrics)
}
\arguments{
\item{metrics}{Wide metric data from \code{compute_core_metrics()}.}
}
\value{
Long-format user-facing metrics.
}
\description{
Prepare metrics for user-facing display
}
