% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the CGManalyzer2 app}
\usage{
run_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{shiny::shinyApp()}.}
}
\value{
A Shiny application object.
}
\description{
Run the CGManalyzer2 app
}
