% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{run_cgmissingdata_imputation}
\alias{run_cgmissingdata_imputation}
\title{Run CGMissingDataR missing glucose imputation}
\usage{
run_cgmissingdata_imputation(data, model = "mice_only", seed = 42)
}
\arguments{
\item{data}{Standardized CGM data.}

\item{model}{Imputation model. Defaults to \code{mice_only}.}

\item{seed}{Integer seed.}
}
\value{
Output from \code{CGMissingDataR::run_missing_glucose_imputation()}.
}
\description{
Run CGMissingDataR missing glucose imputation
}
