% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_adapters.R
\name{to_iglu_data}
\alias{to_iglu_data}
\title{Convert standardized CGM data for iglu}
\usage{
to_iglu_data(data)
}
\arguments{
\item{data}{Standardized CGM data.}
}
\value{
A data frame with \code{id}, \code{time}, and \code{gl} columns.
}
\description{
Convert standardized CGM data for iglu
}
